+++ src/readabs/__init__.py
@@ -1,5 +1,4 @@
 from . import readabs
 from .readabs import *
 
-__version__ = "0.0.7"
-
+__version__ = "0.0.8"